Refactor notifications actions, reducers, and state.
